+++ init.R
@@ -150,7 +150,7 @@ sourceObsmonFiles <- function() {
   source("plots_maps.R")
   source("plots_diagnostic.R")
   source("windspeed.R")
-  source("plots_quick.R")
+  source("plots_multi.R")
 }
 
 fillInDefault <- function(config, key, default) {